# docs/swagger.yaml
basePath: /api
definitions:
  api.JsonError:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  service.Rating:
    properties:
      created_at:
        example: "2025-10-30T00:00:00Z"
        type: string
      movie_id:
        example: 2
        type: integer
      rating:
        example: 5
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  service.RatingData:
    properties:
      movie_id:
        type: integer
      rating:
        maximum: 10
        type: integer
    required:
    - movie_id
    - rating
    type: object
...
  service.UpdatableRatingData:
    properties:
      rating:
        type: integer
    required:
    - rating
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Esta es la API REST para la aplicación de gestión de películas.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API de Películas UNICEN
  version: "1.0"
paths:
...
  /movies/{movieID}/ratings:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de ratings asociados a una película dada por ID,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de ratings exitosa
          schema:
            items:
              $ref: '#/definitions/service.Rating'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista los ratings de una película paginados
      tags:
      - Ratings
...
  /users/{userID}/ratings:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de ratings asociados a un usuario dada por ID,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de ratings exitosa
          schema:
            items:
              $ref: '#/definitions/service.Rating'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista los ratings de un usuario paginados
      tags:
      - Ratings
    post:
      consumes:
      - application/json
      description: Agrega un rating de un usuario dado por ID a una película también
        dada por ID.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: Datos del rating
        in: body
        name: rating
        required: true
        schema:
          $ref: '#/definitions/service.RatingData'
      produces:
      - application/json
      responses:
        "201":
          description: Rating agregada exitosamente
          schema:
            $ref: '#/definitions/service.Rating'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por rating no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicación de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Agrega un rating de un usuario sobre una película
      tags:
      - Ratings
  /users/{userID}/ratings/{movieID}:
    delete:
      consumes:
      - application/json
      description: Borra un rating de un usuario dado por ID sobre una película dada
        también por ID permanentemente.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Rating borrada exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por rating no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Borra un rating de un usuario sobre una película
      tags:
      - Ratings
    get:
      consumes:
      - application/json
      description: Obtiene un rating de un usuario dado por ID sobre una película
        también dada por ID.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Rating obtenida exitosamente
          schema:
            $ref: '#/definitions/service.Rating'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por rating no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Obtiene un rating de un usuario sobre una película
      tags:
      - Ratings
    put:
      consumes:
      - application/json
      description: Actualiza un rating de un usuario dado por ID sobre una película
        dada también por ID.
      parameters:
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: integer
      - description: ID de la película
        in: path
        name: movieID
        required: true
        type: integer
      - description: Datos actualizables del rating
        in: body
        name: rating
        required: true
        schema:
          $ref: '#/definitions/service.UpdatableRatingData'
      produces:
      - application/json
      responses:
        "200":
          description: Rating actualizada exitosamente
          schema:
            $ref: '#/definitions/service.Rating'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por rating no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Actualiza un rating de un usuario sobre una película
      tags:
      - Ratings
...
swagger: "2.0"